api/requirements.txt
fastapi
uvicorn
youtube-transcript-api
google-generativeai
resend
supabase
python-dotenv
pydantic
pydantic-settings
pytest
httpx
ruff
